refactor: move common completions code to its own module
